User Interface

UnivariateUnimodalHighestDensityRegion.univariate_unimodal_HDR

UnivariateUnimodalHighestDensityRegion.univariate_unimodal_HDR — Function

univariate_unimodal_HDR(d::UnivariateDistribution, region::Real, num_steps::Int)

A grid-based approach to find the highest density region (HDR) of a univariate, unimodal distribution. Returns the found HDR interval as a vector with length two.

Arguments

d: a UnivariateDistribution defined in Distributions.jl.

region: a real number ∈ [0, 1] specifying the proportion of the density of d of the highest density region.

num_steps: the number of steps to linearly space between the left and right of the bracket, inclusive, where l=0.0 and r=1.0-region. The lower and upper quantiles, region apart, at each of these steps is then calculated. The the lower and upper quantile at these steps with the minimum distance between them is estimated to be the HDR. num_steps is enforced to be at least 3 and odd to allow exact accuracy on symmetric distributions.

univariate_unimodal_HDR(d::UnivariateDistribution, 
    region::Real,
    solve_alg; 
    solve_kwargs::NamedTuple=(reltol = 0.0001,))

This is an optimization-based approach to find the highest density region (HDR) of a univariate, unimodal distribution. The grid-based approach may be more robust for discrete distributions. Returns the found HDR interval as a vector with length two.

Requires loading Optimization.jl and an additional package that contains optimization algorithms: OptimizationNLopt.jl may be a good starting point.

Arguments

d: a UnivariateDistribution defined in Distributions.jl.

region: a real number ∈ [0, 1] specifying the proportion of the density of d of the highest density region.

solve_alg: an algorithm to use to solve for the highest density region defined within Optimization.jl. Good starting methods from NLopt may be NLopt.LN_BOBYQA(), NLopt.LN_NELDERMEAD() and NLopt.LD_LBFGS. Other packages can be used as well - see Overview of the Optimizers.

Keyword Arguments

solve_kwargs: a NamedTuple of keyword arguments used to set solver options like reltol. Other arguments can be used, but the default of (reltol = 0.0001,) is expected to be sufficient. For a list of common solver arguments see: Common Solver Options. Other specific package arguments may also be available. For NLopt see Methods.